requirements.txt
Flask~=3.1.0
python-dotenv~=1.0.1
Flask-Migrate~=4.1.0
Werkzeug~=3.1.3
WTForms~=3.2.1
Flask-Login~=0.6.3
face-recognition~=1.3.0
numpy~=2.2.3
opencv-python~=4.11.0.86
pillow~=11.1.0
pytz~=2025.1